requirements.txt
# FastAPI Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Machine Learning & Computer Vision
tensorflow==2.15.0
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pyaudio==0.2.11

# Music & Spotify Integration
spotipy==2.22.1
requests==2.31.0

# Database & Caching
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
alembic==1.13.1

# Image Processing
Pillow==10.1.0
matplotlib==3.8.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
httpx==0.25.2
aiofiles==23.2.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Deployment
gunicorn==21.2.0
docker==6.1.3
